piece is removed if another moves to its square
